Updated passedMetFilters function.
The branch "Flag_ecalBadCalibFilterV2" does not exist in 2016, and we don't need it for this year. Therefore, I'll only set the branch in case we are analyzing 2017 or 2018 data. 
IMPORTANT NOTE: In case you are running MC, it's important to still set the JSONFileName corresponding to the year you are analyzing because this function takes this name as an argument. So make sure that this option corresponds to the year you are analyzing when running both data and MC.