Added 3 options for b-tag SFs (important)
We need to specify in lowercase the working point of b-tag ID we are using (loose, medium, tight) in JetBTaggingWP. This only concerns to the b-tagging scale factor part! So you still need to change manually the b-tagging discriminator threshold in JetBTaggingCut. We also have a new option "BtagSFCSVfile" where we will put the name of the CSV file that contains the info of the SF for the year/b-tag algorithm we are using for our analysis. These CSV files must be placed in the Pileup directory.