Adding a functionality that derives specific PU wgts per MC sample
This functionality is added in the Analyzer constructor. If specialPUcalculation is true, then take the name of the output file (when submitting jobs) to retrieve the right MC nTruePU distribution to calculate the PU weights, otherwise, it will do the calculation as usual, having a single MC nTruePU histogram. If you need to run the Analyzer interactively and use the specialPUcalculation option, make sure that you include the name of the sample to analyze. For example, you can look at the names given to output files in past runs submitted as jobs and use the convention of the output filenames there. This will be improved later.