New options: SpecialMCPUCalculation + SpecialMCPUHistos
These options are similar to MCHistos (root file that contains the PU distribution for all MC samples) and MCPUHistName (name of this histogram in the root file specified). The difference is now that there is a histogram named according to the sample analyzed that will be set once the Analyzer is running. Here we just need to turn on/off the specialMCPUcalculation option and give the name of the root file that contains the nTruePU histograms of the samples we are analyzing. 

***IMPORTANT NOTES***:

- When running over data, ***make sure that the SpecialMCPUCalculation option is set to false***. Otherwise, you will get failed jobs (this will be fixed/improved later).
- The root files with the PU distributions should be stored in the Pileup directory.
- Check the histograms stored in each root file and make sure that the samples you are using are included.
- For convenience, I'm leaving there all the files that are available right now for 2017 MC:
    * new_mc2017_pileupReweighting_NanoAODv6.root : most of the common MC samples, except the W+Jets and DY+Jets inclusive ones.
    * new_mc2017_pileupReweighting_NanoAODv6_wanddylo_incl.root : W+Jets and DY+Jets LO (leading order) inclusive samples distributions.
    * new_mc2017_pileupReweighting_NanoAODv6_dynlo_incl.root: DY+Jets NLO (next-to-leading order) inclusive sample distribution.